+++ src/AuthCode.php
@@ -20,4 +20,14 @@ class AuthCode extends Model
     {
         return $this->referencesOne(Client::class, 'client_id');
     }
+
+    /**
+     * Get the current connection name for the model.
+     *
+     * @return string|null
+     */
+    public function getConnectionName()
+    {
+        return config('passport.storage.database.connection') ?? $this->connection;
+    }
 }