+++ src/ClientRepository.php
@@ -16,7 +16,9 @@ class ClientRepository
      */
     public function find($id)
     {
-        return Client::first($id);
+        $clientModel = Passport::clientModel();
+
+        return $clientModel::first($id);
     }
 
     /**
@@ -30,7 +32,7 @@ class ClientRepository
     {
         $client = $this->find($id);
 
-        return $client && !$client->revoked ? $client : null;
+        return $client && ! $client->revoked ? $client : null;
     }
 
     /**
@@ -43,7 +45,9 @@ class ClientRepository
      */
     public function findForUser($clientId, $userId)
     {
-        return Client::first(['_id' => (string) $clientId, 'user_id' => (string) $userId]);
+        $clientModel = Passport::clientModel();
+
+        return $clientModel::first(['_id' => (string) $clientId, 'user_id' => (string) $userId]);
     }
 
     /**
@@ -55,7 +59,9 @@ class ClientRepository
      */
     public function forUser($userId)
     {
-        return Client::where(['user_id' => (string) $userId])
+        $clientModel = Passport::clientModel();
+
+        return $clientModel::where(['user_id' => (string) $userId])
             ->sort(['name' => 1]);
     }
 
@@ -110,6 +116,7 @@ class ClientRepository
      * @param  int    $userId
      * @param  string $name
      * @param  string $redirect
+     * @param  string|null  $provider
      * @param  bool   $personalAccess
      * @param  bool   $password
      * @param  bool   $confidential
@@ -117,7 +124,7 @@ class ClientRepository
      *
      * @return \Laravel\Passport\Client
      */
-    public function create($userId, $name, $redirect, $personalAccess = false, $password = false, $confidential = true, $allowedScopes = null)
+    public function create($userId, $name, $redirect, $provider = null, $personalAccess = false, $password = false, $confidential = true, $allowedScopes = null)
     {
         $client = Passport::client();
 
@@ -126,6 +133,7 @@ class ClientRepository
                 'user_id' => $userId,
                 'name' => $name,
                 'secret' => ($confidential || $personalAccess) ? Str::random(40) : null,
+                'provider' => $provider,
                 'redirect' => $redirect,
                 'personal_access_client' => $personalAccess,
                 'password_client' => $password,
@@ -151,7 +159,7 @@ class ClientRepository
      */
     public function createPersonalAccessClient($userId, $name, $redirect)
     {
-        return tap($this->create($userId, $name, $redirect, true), function ($client) {
+        return tap($this->create($userId, $name, $redirect, null, true), function ($client) {
             $accessClient = Passport::personalAccessClient();
             $accessClient->client_id = $client->_id;
             $accessClient->save();
@@ -164,12 +172,13 @@ class ClientRepository
      * @param  int    $userId
      * @param  string $name
      * @param  string $redirect
+     * @param  string|null  $provider
      *
      * @return \Laravel\Passport\Client
      */
-    public function createPasswordGrantClient($userId, $name, $redirect)
+    public function createPasswordGrantClient($userId, $name, $redirect, $provider = null)
     {
-        return $this->create($userId, $name, $redirect, false, true);
+        return $this->create($userId, $name, $redirect, $provider, false, true);
     }
 
     /**