+++ src/Console/ClientCommand.php
@@ -2,10 +2,10 @@
 
 namespace Laravel\Passport\Console;
 
-use Laravel\Passport\Passport;
 use Illuminate\Console\Command;
 use Laravel\Passport\Client;
 use Laravel\Passport\ClientRepository;
+use Laravel\Passport\Passport;
 
 class ClientCommand extends Command
 {
@@ -19,6 +19,7 @@ class ClientCommand extends Command
             {--password : Create a password grant client}
             {--client : Create a client credentials grant client}
             {--name= : The name of the client}
+            {--provider= : The name of the user provider}
             {--redirect_uri= : The URI to redirect to after authorization }
             {--user_id= : The user ID the client should be assigned to }
             {--public : Create a public client (Auth code grant type only) }';
@@ -84,8 +85,16 @@ class ClientCommand extends Command
             config('app.name').' Password Grant Client'
         );
 
+        $providers = array_keys(config('auth.providers'));
+
+        $provider = $this->option('provider') ?: $this->choice(
+            'Which user provider should this client use to retrieve users?',
+            $providers,
+            in_array('users', $providers) ? 'users' : null
+        );
+
         $client = $clients->createPasswordGrantClient(
-            null, $name, 'http://localhost'
+            null, $name, 'http://localhost', $provider
         );
 
         $this->info('Password grant client created successfully.');
@@ -149,7 +158,7 @@ class ClientCommand extends Command
         $allowedScopes = config('auth.authorization_code.allowed_scopes');
 
         $client = $clients->create(
-            $userId, $name, $redirect, false, false, ! $this->option('public'), $allowedScopes
+            $userId, $name, $redirect, null, false, false, ! $this->option('public'), $allowedScopes
         );
 
         $this->info('New client created successfully.');
@@ -165,8 +174,13 @@ class ClientCommand extends Command
      */
     protected function outputClientDetails(Client $client)
     {
+        if (Passport::$hashesClientSecrets) {
+            $this->line('<comment>Here is your new client secret. This is the only time it will be shown so don\'t lose it!</comment>');
+            $this->line('');
+        }
+
         $this->line('<comment>Client ID:</comment> '.$client->_id);
-        $this->line('<comment>Client secret:</comment> '.$client->secret);
+        $this->line('<comment>Client secret:</comment> '.$client->plainSecret);
     }
 
     /**