+++ src/Http/Controllers/PersonalAccessTokenController.php
@@ -47,11 +47,11 @@ class PersonalAccessTokenController
      */
     public function forUser(Request $request)
     {
-        $tokens = $this->tokenRepository->forUser($request->user()->getKey());
+        $tokens = $this->tokenRepository->forUser($request->user()->getAuthIdentifier());
 
         return collect($tokens)->filter(
             function ($token) {
-                return !$token->revoked && $token->client()->personal_access_client;
+                return $token->client()->personal_access_client && ! $token->revoked;
             }
         )->values();
     }
@@ -65,13 +65,10 @@ class PersonalAccessTokenController
      */
     public function store(Request $request)
     {
-        $this->validation->make(
-            $request->all(),
-            [
-                'name' => 'required|max:255',
-                'scopes' => 'array|in:'.implode(',', Passport::scopeIds()),
-            ]
-        )->validate();
+        $this->validation->make($request->all(), [
+            'name' => 'required|max:191',
+            'scopes' => 'array|in:'.implode(',', Passport::scopeIds()),
+        ])->validate();
 
         return $request->user()->createToken(
             $request->name,
@@ -89,8 +86,7 @@ class PersonalAccessTokenController
     public function destroy(Request $request, $tokenId)
     {
         $token = $this->tokenRepository->findForUser(
-            $tokenId,
-            $request->user()->getKey()
+            $tokenId, $request->user()->getAuthIdentifier()
         );
 
         if (is_null($token)) {